TST: fixed up indexing
