+++ pandas/core/frame.py
@@ -2799,233 +2799,6 @@ class DataFrame(NDFrame):
         return result
 
     #----------------------------------------------------------------------
-    # Filling NA's
-
-    def fillna(self, value=None, method=None, axis=0, inplace=False,
-               limit=None, downcast=None):
-        """
-        Fill NA/NaN values using the specified method
-
-        Parameters
-        ----------
-        method : {'backfill', 'bfill', 'pad', 'ffill', None}, default None
-            Method to use for filling holes in reindexed Series
-            pad / ffill: propagate last valid observation forward to next valid
-            backfill / bfill: use NEXT valid observation to fill gap
-        value : scalar or dict
-            Value to use to fill holes (e.g. 0), alternately a dict of values
-            specifying which value to use for each column (columns not in the
-            dict will not be filled)
-        axis : {0, 1}, default 0
-            0: fill column-by-column
-            1: fill row-by-row
-        inplace : boolean, default False
-            If True, fill the DataFrame in place. Note: this will modify any
-            other views on this DataFrame, like if you took a no-copy slice of
-            an existing DataFrame, for example a column in a DataFrame. Returns
-            a reference to the filled object, which is self if inplace=True
-        limit : int, default None
-            Maximum size gap to forward or backward fill
-        downcast : dict, default is None, a dict of item->dtype of what to 
-            downcast if possible
-
-        See also
-        --------
-        reindex, asfreq
-
-        Returns
-        -------
-        filled : DataFrame
-        """
-        self._consolidate_inplace()
-
-        axis = self._get_axis_number(axis)
-        if value is None:
-            if method is None:
-                raise ValueError('must specify a fill method or value')
-            if self._is_mixed_type and axis == 1:
-                if inplace:
-                    raise NotImplementedError()
-                return self.T.fillna(method=method, limit=limit).T
-
-            method = com._clean_fill_method(method)
-            new_data = self._data.interpolate(method  = method,
-                                              axis    = axis,
-                                              limit   = limit,
-                                              inplace = inplace,
-                                              coerce  = True)
-        else:
-            if method is not None:
-                raise ValueError('cannot specify both a fill method and value')
-            # Float type values
-            if len(self.columns) == 0:
-                return self
-            if isinstance(value, (dict, Series)):
-                if axis == 1:
-                    raise NotImplementedError('Currently only can fill '
-                                              'with dict/Series column '
-                                              'by column')
-
-                result = self if inplace else self.copy()
-                for k, v in value.iteritems():
-                    if k not in result:
-                        continue
-                    result[k].fillna(v, inplace=True)
-                return result
-            else:
-                new_data = self._data.fillna(value, inplace=inplace, 
-                                             downcast=downcast)
-
-        if inplace:
-            self._data = new_data
-        else:
-            return self._constructor(new_data)
-
-    def ffill(self, axis=0, inplace=False, limit=None):
-        return self.fillna(method='ffill', axis=axis, inplace=inplace,
-                           limit=limit)
-
-    def bfill(self, axis=0, inplace=False, limit=None):
-        return self.fillna(method='bfill', axis=axis, inplace=inplace,
-                           limit=limit)
-
-    def replace(self, to_replace, value=None, method='pad', axis=0,
-                inplace=False, limit=None):
-        """
-        Replace values given in 'to_replace' with 'value' or using 'method'
-
-        Parameters
-        ----------
-        value : scalar or dict, default None
-            Value to use to fill holes (e.g. 0), alternately a dict of values
-            specifying which value to use for each column (columns not in the
-            dict will not be filled)
-        method : {'backfill', 'bfill', 'pad', 'ffill', None}, default 'pad'
-            Method to use for filling holes in reindexed Series
-            pad / ffill: propagate last valid observation forward to next valid
-            backfill / bfill: use NEXT valid observation to fill gap
-        axis : {0, 1}, default 0
-            0: fill column-by-column
-            1: fill row-by-row
-        inplace : boolean, default False
-            If True, fill the DataFrame in place. Note: this will modify any
-            other views on this DataFrame, like if you took a no-copy slice of
-            an existing DataFrame, for example a column in a DataFrame. Returns
-            a reference to the filled object, which is self if inplace=True
-        limit : int, default None
-            Maximum size gap to forward or backward fill
-
-        See also
-        --------
-        reindex, asfreq
-
-        Returns
-        -------
-        filled : DataFrame
-        """
-        self._consolidate_inplace()
-
-        axis = self._get_axis_number(axis)
-
-        if value is None:
-            return self._interpolate(to_replace, method, axis, inplace, limit)
-        else:
-            if len(self.columns) == 0:
-                return self
-
-            new_data = self._data
-            if isinstance(to_replace, (dict, Series)):
-                if isinstance(value, (dict, Series)):  # {'A' : NA} -> {'A' : 0}
-                    new_data = self._data
-                    for c, src in to_replace.iteritems():
-                        if c in value and c in self:
-                            new_data = new_data.replace(src, value[c],
-                                                        filter=[ c ],
-                                                        inplace=inplace)
-
-                elif not isinstance(value, (list, np.ndarray)):
-                    new_data = self._data
-                    for k, src in to_replace.iteritems():
-                        if k in self:
-                            new_data = new_data.replace(src, value,
-                                                        filter = [ k ],
-                                                        inplace=inplace)
-                else:
-                    raise ValueError('Fill value must be scalar or dict or Series')
-
-            elif isinstance(to_replace, (list, np.ndarray)):
-                # [NA, ''] -> [0, 'missing']
-                if isinstance(value, (list, np.ndarray)):
-                    if len(to_replace) != len(value):
-                        raise ValueError('Replacement lists must match '
-                                         'in length. Expecting %d got %d ' %
-                                         (len(to_replace), len(value)))
-
-                    new_data = self._data.replace_list(to_replace, value,
-                                                       inplace=inplace)
-
-                else:  # [NA, ''] -> 0
-                    new_data = self._data.replace(to_replace, value,
-                                                  inplace=inplace)
-
-            else:
-
-                # dest iterable dict-like
-                if isinstance(value, (dict, Series)):  # NA -> {'A' : 0, 'B' : -1}
-
-                    new_data = self._data
-                    for k, v in value.iteritems():
-                        if k in self:
-                            new_data = new_data.replace(to_replace, v,
-                                                        filter=[ k ],
-                                                        inplace=inplace)
-
-                elif not isinstance(value, (list, np.ndarray)):  # NA -> 0
-                    new_data = self._data.replace(to_replace, value,
-                                                  inplace=inplace)
-                else:
-                    raise ValueError('Invalid to_replace type: %s' %
-                                     type(to_replace))  # pragma: no cover
-
-
-        if inplace:
-            self._data = new_data
-        else:
-            return self._constructor(new_data)
-
-    def _interpolate(self, to_replace, method, axis, inplace, limit):
-        if self._is_mixed_type and axis == 1:
-            return self.T.replace(to_replace, method=method, limit=limit).T
-
-        method = com._clean_fill_method(method)
-
-        if isinstance(to_replace, (dict, Series)):
-            if axis == 1:
-                return self.T.replace(to_replace, method=method,
-                                      limit=limit).T
-
-            rs = self if inplace else self.copy()
-            for k, v in to_replace.iteritems():
-                if k in rs:
-                    rs[k].replace(v, method=method, limit=limit,
-                                  inplace=True)
-            return rs if not inplace else None
-
-        else:
-
-            new_data = self._data.interpolate(method  = method,
-                                              axis    = axis,
-                                              limit   = limit,
-                                              inplace = inplace,
-                                              missing = to_replace,
-                                              coerce  = False)
-
-            if inplace:
-                self._data = new_data
-            else:
-                return self._constructor(new_data)
-
-    #----------------------------------------------------------------------
     # Rename
 
     def rename(self, index=None, columns=None, copy=True, inplace=False):
